upload page w/ essential components
Moved the file uploader and video recorder to the same page, added a functioning progress bar. Imported chakra ui
