+++ pages/index.js
@@ -4,19 +4,22 @@ import styles from '../styles/Home.module.css'
 import React from 'react'
 import ReactPlayer from 'react-player'
 import firebase from '../firebase/clientApp'
+import { Flex, Spacer } from "@chakra-ui/react"
 
 export const Home = ({ videoURL }) => {
   return (
+    <Flex alignItems='center'>
     <div>
       <ReactPlayer url={ videoURL } playing={true} loop={true} />
     </div>
+    </Flex>
   );
 };
 
 export const getServerSideProps = async () => {
   const db = firebase.firestore();
   // Create a reference to a document in the videos collection
-  var docRef = db.collection("videos").doc("nBbL35I9MXVZu0vsq7LL");
+  var docRef = db.collection("videos").doc("WeSIVSdvukD29ZhiFLPQ");
   var videoURL;
   // Retrieve the video URL from the document
   await docRef.get().then((doc) => {
